requirements.txt
Django==1.6.6
boto==2.32.1
dj-database-url==0.3.0
django-heroku-memcacheify==0.5
django-heroku-postgresify==0.3
django-herokuify==1.0.pre3
django-pylibmc-sasl==0.2.4
django-storages==1.1.8
django-tastypie==0.11.1
pylibmc==1.3.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
wsgiref==0.1.2
